AddedEventCreationFunctionality
Finished Create Event, and added a method to get Event by employeeID.
Added Splash screen layout.

NOTES: Adapter updating is not work at the moment. will have to refactor
refresh methods.
Need some way to keep track of the employeeID of the employee that is
logged in.
Need to fix database to auto increment EventIDS instead of setting it